feat: fixed layouts
